Update tests style
